+++ term_image/image/common.py
@@ -14,6 +14,7 @@ import time
 from abc import ABC, abstractmethod
 from enum import Enum
 from functools import wraps
+from math import ceil
 from operator import gt, mul, sub
 from random import randint
 from types import FunctionType, TracebackType
@@ -31,7 +32,7 @@ from ..exceptions import (
     URLNotFoundError,
     _style_error,
 )
-from ..utils import ClassInstanceMethod, get_terminal_size, no_redecorate
+from ..utils import ClassInstanceMethod, get_cell_size, get_terminal_size, no_redecorate
 
 _ALPHA_THRESHOLD = 40 / 255  # Default alpha threshold
 _FORMAT_SPEC = re.compile(
@@ -499,7 +500,7 @@ class BaseImage(ABC):
         pad_width: Optional[int] = None,
         v_align: Optional[str] = None,
         pad_height: Optional[int] = None,
-        alpha: Optional[float] = _ALPHA_THRESHOLD,
+        alpha: Optional[float, str] = _ALPHA_THRESHOLD,
         *,
         scroll: bool = False,
         animate: bool = True,
@@ -586,9 +587,10 @@ class BaseImage(ABC):
 
           * *scroll* is ignored.
           * Image size and :term:`padding height` are always validated, if set or given.
+          * **with the exception of native animations provided by some render styles**.
 
         * Animations, **by default**, are infinitely looped and can be terminated
-          with ``Ctrl-C`` (``SIGINT``), raising ``KeyboardInterrupt``.
+          with **Ctrl+C** (``SIGINT``), raising ``KeyboardInterrupt``.
         """
         fmt = self._check_formatting(h_align, pad_width, v_align, pad_height)
 
@@ -614,7 +616,8 @@ class BaseImage(ABC):
             )
 
         if (
-            self._is_animated
+            not style.get("native")
+            and self._is_animated
             and animate
             and None is not pad_height > get_terminal_size()[1]
         ):
@@ -659,7 +662,7 @@ class BaseImage(ABC):
             render,
             scroll=scroll,
             check_size=check_size,
-            animated=self._is_animated and animate,
+            animated=not style.get("native") and self._is_animated and animate,
         )
 
     @classmethod
@@ -1731,6 +1734,48 @@ class GraphicsImage(BaseImage):
                 "This image render style is not supported in the active terminal"
             )
         super().__init__(image, **kwargs)
+
+    def _get_minimal_render_size(self) -> Tuple[int, int]:
+        render_size = self._get_render_size()
+        r_height = self.rendered_height
+        width, height = (
+            render_size
+            if mul(*render_size) < mul(*self._original_size)
+            else self._original_size
+        )
+
+        # When `_original_size` is used, ensure the height is a multiple of the rendered
+        # height, so that pixels can be evenly distributed among all lines.
+        # If r_height == 0, height == 0, extra == 0; Handled in `_get_render_data()`.
+        extra = height % (r_height or 1)
+        if extra:
+            # Incremented to the greater multiple to avoid losing any data
+            height = height - extra + r_height
+
+        return width, height
+
+    def _get_render_size(self) -> Tuple[int, int]:
+        return tuple(map(mul, self.rendered_size, get_cell_size() or (1, 2)))
+
+    @staticmethod
+    def _pixels_cols(
+        *, pixels: Optional[int] = None, cols: Optional[int] = None
+    ) -> int:
+        return (
+            ceil(pixels // (get_cell_size() or (1, 2))[0])
+            if pixels is not None
+            else cols * (get_cell_size() or (1, 2))[0]
+        )
+
+    @staticmethod
+    def _pixels_lines(
+        *, pixels: Optional[int] = None, lines: Optional[int] = None
+    ) -> int:
+        return (
+            ceil(pixels // (get_cell_size() or (1, 2))[1])
+            if pixels is not None
+            else lines * (get_cell_size() or (1, 2))[1]
+        )
 
 
 class TextImage(BaseImage):